+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Support `$string_split`
+- Support named `item` variable in foreach
+
 ## [2.0.2] - 2021-12-20
 
 ### Fixed